Move buildFlagWithPopover to BasePage so all pages can use it.  And use on the Instructor and Student grade drilldown pages (both modal and student views)
