+++ tool/src/java/org/sakaiproject/gradebookng/tool/pages/GradebookPage.java
@@ -500,20 +500,4 @@ public class GradebookPage extends BasePage {
 	}
 	
 	
-	/**
-	 * Helper to build a notification flag with a Bootstrap popover 
-	 */
-	public WebMarkupContainer buildFlagWithPopover(String componentId, String popoverContent) {
-		WebMarkupContainer flagWithPopover = new WebMarkupContainer(componentId);
-
-		flagWithPopover.add(new AttributeModifier("data-toggle", "popover"));
-		flagWithPopover.add(new AttributeModifier("data-trigger", "focus"));
-		flagWithPopover.add(new AttributeModifier("data-placement", "bottom"));
-		flagWithPopover.add(new AttributeModifier("data-html", "true"));
-		//flagWithPopover.add(new AttributeModifier("data-container", "#gradebookGrades"));
-		flagWithPopover.add(new AttributeModifier("data-content", popoverContent));
-		flagWithPopover.add(new AttributeModifier("tabindex", "0"));
-
-		return flagWithPopover;
-	}
 }
