+++ tool/src/java/org/sakaiproject/gradebookng/tool/panels/InstructorGradeSummaryGradesPanel.html
@@ -45,9 +45,9 @@
 					<td>
 						<span class="gb-summary-grade-title" wicket:id="title"></span>
 						<span class="gb-summary-grade-flags" wicket:id="flags">
-							<span wicket:id="isExtraCredit" class="gb-flag-extra-credit" wicket:message="title:label.gradeitem.extracredit"></span>
-							<span wicket:id="isNotCounted" class="gb-flag-not-counted" wicket:message="title:label.gradeitem.notcounted"></span>
-							<span wicket:id="isNotReleased" class="gb-flag-not-released" wicket:message="title:label.gradeitem.notreleased"></span>
+							<span wicket:id="isExtraCredit" class="gb-flag-extra-credit"></span>
+							<span wicket:id="isNotCounted" class="gb-flag-not-counted"></span>
+							<span wicket:id="isNotReleased" class="gb-flag-not-released"></span>
 						</span>
 					</td>
 					<td class="gb-summary-grade-score">
